requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.8.0
botocore==1.12.228
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
docutils==0.15.2
future==0.17.1
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
pynamodb==4.0.0
python-dateutil==2.8.0
python-telegram-bot==12.0.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
